requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.6
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-extensions==3.1.5
django-split-settings==1.1.0
django-widget-tweaks==1.4.12
executing==0.8.3
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==5.0.0
graphviz==0.20
gunicorn==20.1.0
hiredis==2.0.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
kombu==5.2.4
Markdown==3.4.1
matplotlib-inline==0.1.3
mccabe==0.6.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.12.0
pygraphviz==1.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytils==0.4.1
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
